refator: implement FeedbackRow component and integrate with FeedbackTable
